Made changes to cmdlets

+++ src/EventHub/EventHub/Cmdlets/SchemaRegistry/RemoveAzureEventHubSchemaGroup.cs
@@ -4,24 +4,34 @@ using System.Text;
 using Microsoft.Azure.Commands.EventHub.Models;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using System.Management.Automation;
+using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 
 namespace Microsoft.Azure.Commands.EventHub.Commands.SchemaRegistry
 {
     [Cmdlet("Remove", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventHubSchemaGroup", SupportsShouldProcess = true), OutputType(typeof(PSEventHubsSchemaRegistryAttributes))]
     public class RemoveAzureEventHubSchemaGroup: AzureEventHubsCmdletBase
     {
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, Position = 0, HelpMessage = "Resource Group Name")]
+        [Parameter(Mandatory = true, ParameterSetName = NamespaceSchemaGroupParameterSet, ValueFromPipelineByPropertyName = true, Position = 0, HelpMessage = "Resource Group Name")]
         [ResourceGroupCompleter]
         [ValidateNotNullOrEmpty]
         public string ResourceGroupName { get; set; }
 
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, Position = 1, HelpMessage = "Namespace Name")]
+        [Parameter(Mandatory = true, ParameterSetName = NamespaceSchemaGroupParameterSet, ValueFromPipelineByPropertyName = true, Position = 1, HelpMessage = "Namespace Name")]
         [Alias(AliasNamespaceName)]
         public string Namespace { get; set; }
 
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, Position = 2, HelpMessage = "Name of Schema Group")]
+        [Parameter(Mandatory = true, ParameterSetName = NamespaceSchemaGroupParameterSet, ValueFromPipelineByPropertyName = true, Position = 2, HelpMessage = "Name of Schema Group")]
+        [Parameter(Mandatory = true, ParameterSetName = SchemaGroupResourceIdParameterSet, ValueFromPipelineByPropertyName = true, Position = 2, HelpMessage = "Name of Schema Group")]
         [Alias(AliasSchemaGroupName)]
         public string Name { get; set; }
+
+        [Parameter(Mandatory = true, ParameterSetName = SchemaGroupInputObjectParameterSet, ValueFromPipeline = true, Position = 0, HelpMessage = "Namespace Object")]
+        [ValidateNotNullOrEmpty]
+        public PSEventHubsSchemaRegistryAttributes InputObject { get; set; }
+
+        [Parameter(Mandatory = true, ParameterSetName = SchemaGroupResourceIdParameterSet, ValueFromPipelineByPropertyName = true, Position = 0, HelpMessage = "Namespace Resource Id")]
+        [ValidateNotNullOrEmpty]
+        public string ResourceId { get; set; }
 
         [Parameter(Mandatory = false)]
         public SwitchParameter PassThru { get; set; }
@@ -32,11 +42,40 @@ namespace Microsoft.Azure.Commands.EventHub.Commands.SchemaRegistry
             {
                 try
                 {
-                    Client.DeleteNamespaceSchemaGroup(ResourceGroupName, Namespace, Name);
-                    if (PassThru)
+
+                    if (ParameterSetName == SchemaGroupInputObjectParameterSet)
                     {
-                        WriteObject(true);
+                        LocalResourceIdentifier getParamSchemaGroup = new LocalResourceIdentifier(InputObject.Id);
+                        ResourceGroupName = getParamSchemaGroup.ResourceGroupName;
+                        Namespace = getParamSchemaGroup.ParentResource;
+                        Name = getParamSchemaGroup.ResourceName;
                     }
+
+                    if(ParameterSetName == SchemaGroupResourceIdParameterSet)
+                    {
+                        LocalResourceIdentifier getParamSchemaGroup = new LocalResourceIdentifier(ResourceId);
+                        ResourceGroupName = getParamSchemaGroup.ResourceGroupName;
+                        Namespace = getParamSchemaGroup.ParentResource;
+                        Name = getParamSchemaGroup.ResourceName;
+                    }
+
+                    if (ShouldProcess(target: Name, action: string.Format(Resources.RemoveNamespacesSchemaGroup, Name, Namespace)))
+                    {
+                        try
+                        {
+                            Client.DeleteNamespaceSchemaGroup(ResourceGroupName, Namespace, Name);
+                            if (PassThru)
+                            {
+                                WriteObject(true);
+                            }
+                        }
+                        catch (Management.EventHub.Models.ErrorResponseException ex)
+                        {
+                            WriteError(Eventhub.EventHubsClient.WriteErrorforBadrequest(ex));
+                        }
+                    }
+                        
+                    
                 }
                 catch (Management.EventHub.Models.ErrorResponseException ex)
                 {
